Neutral Point & Static Margin & Vertical CG Calculation and Plotting
Sorry for minor delay on the stability part due to the issue related to the compressible aerodynamics, a empirical way has been decided and carried out to complete the analysis.

Trim code is based on the equation solving with two unknowns, once the parameters being filled in Excel the plot would be updated.